renaming dialogs_service.py --> misc_dialogs.py
