+++ qtodotxt/app.py
@@ -8,7 +8,7 @@ import sys
 from PySide import QtGui
 
 from qtodotxt.ui.controllers.main_controller import MainController
-from qtodotxt.ui.dialogs.dialogs_service import DialogsService
+from qtodotxt.ui.dialogs.misc_dialogs import DialogsService
 from qtodotxt.ui.dialogs.task_editor_service import TaskEditorService
 from qtodotxt.ui.resource_manager import getIcon
 from qtodotxt.ui.views.main_view import MainView
